Replace many unsigned int for size_t. Compiles on VS but not tested.

+++ libhdt/include/Iterator.hpp
@@ -218,13 +218,13 @@ public:
 	virtual bool canGoTo() {
 		return false;
 	}
-	virtual void goTo(unsigned int /*pos*/) {
+	virtual void goTo(size_t /*pos*/) {
 	/* Absolute repositioning of index: ie. go to the index set in the given argument (pos) */
 	}
-	virtual void skip(unsigned int /*pos*/) {
+	virtual void skip(size_t /*pos*/) {
 	/* Relative repositioning of index: ie. skip index by given argument places */
 	}
-	virtual bool findNextOccurrence(unsigned int /*value*/, unsigned char /*component*/) {
+    virtual bool findNextOccurrence(size_t /*value*/, unsigned char /*component*/) {
 		return false;
 	}
 	virtual TripleComponentOrder getOrder() {
